# openapi/prepflow.v1.yaml
openapi: 3.0.3
info:
  title: PrepFlow API
  version: "v1"
  description: |
    PrepFlow backend API (modular monolith).
servers:
  - url: http://localhost:8080

tags:
  - name: Auth
  - name: Users
  - name: Problems
  - name: Reviews
  - name: Lists
  - name: Contests
  - name: Stats
  - name: Calendar

paths:
  /healthz:
    get:
      summary: Liveness probe
      responses:
        "204":
          description: OK
  /readyz:
    get:
      summary: Readiness probe
      responses:
        "204":
          description: Ready
        "503":
          description: Not ready

  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: Invalid request
        "409":
          description: Email already registered
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid credentials
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token and issue a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid refresh token
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Revoke refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "204":
          description: No content
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Unauthorized

  /api/v1/users/me/settings:
    patch:
      tags: [Users]
      summary: Update current user settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchSettingsRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsResponse"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized

  /api/v1/problems/:
    post:
      tags: [Problems]
      summary: Add a problem to the library (creates user state)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProblemRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
    get:
      tags: [Problems]
      summary: List user problems with state
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProblemWithState"
        "401":
          description: Unauthorized

  /api/v1/problems/{id}:
    patch:
      tags: [Problems]
      summary: Update a problem (metadata and/or active state)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchProblemRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "204":
          description: No content
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /api/v1/reviews/due:
    get:
      tags: [Reviews]
      summary: List due items up to a window
      security:
        - bearerAuth: []
      parameters:
        - name: window_days
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProblemWithState"
        "401":
          description: Unauthorized

  /api/v1/reviews/:
    post:
      tags: [Reviews]
      summary: Record a review attempt and reschedule
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostReviewRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostReviewResponse"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /api/v1/integrations/calendar/ics/rotate:
    post:
      tags: [Calendar]
      summary: Rotate calendar subscription token and return a subscription URL
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RotateICSResponse"
        "401":
          description: Unauthorized

  /api/v1/integrations/calendar/ics:
    get:
      tags: [Calendar]
      summary: ICS feed (public via token)
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: window_days
          in: query
          required: false
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: OK
          content:
            text/calendar:
              schema:
                type: string
        "404":
          description: Not found

  /api/v1/lists/:
    post:
      tags: [Lists]
      summary: Create a custom list
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/List"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
    get:
      tags: [Lists]
      summary: List user lists
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/List"
        "401":
          description: Unauthorized

  /api/v1/lists/import:
    post:
      tags: [Lists]
      summary: Import a template list (snapshot copy)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportListRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListWithItems"
        "401":
          description: Unauthorized
        "404":
          description: Template not found

  /api/v1/lists/{id}:
    get:
      tags: [Lists]
      summary: Get list details and items
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListWithItems"
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /api/v1/lists/{id}/items:
    post:
      tags: [Lists]
      summary: Add an existing problem to a list
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddListItemRequest"
      responses:
        "204":
          description: No content
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /api/v1/lists/{id}/items/reorder:
    patch:
      tags: [Lists]
      summary: Reorder list items by problem_id
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderListRequest"
      responses:
        "204":
          description: No content
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /api/v1/contests/generate:
    post:
      tags: [Contests]
      summary: Generate a contest from user problems
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateContestRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContestWithItems"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized

  /api/v1/contests/{id}:
    get:
      tags: [Contests]
      summary: Get contest details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContestWithItems"
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /api/v1/contests/{id}/start:
    post:
      tags: [Contests]
      summary: Start a contest (sets started_at)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contest"
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /api/v1/contests/{id}/complete:
    post:
      tags: [Contests]
      summary: Complete a contest (sets completed_at)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contest"
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /api/v1/contests/{id}/results:
    post:
      tags: [Contests]
      summary: Submit contest results (writes review logs and updates scheduling)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitContestResultsRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContestWithItems"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /api/v1/stats/overview:
    get:
      tags: [Stats]
      summary: Get overview stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsOverview"
        "401":
          description: Unauthorized

  /api/v1/stats/topics:
    get:
      tags: [Stats]
      summary: Topic mastery summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopicStat"
        "401":
          description: Unauthorized

  /api/v1/stats/streaks:
    get:
      tags: [Stats]
      summary: Review streak stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreaksResponse"
        "401":
          description: Unauthorized

  /api/v1/stats/contests:
    get:
      tags: [Stats]
      summary: Contest stats (based on contest results)
      security:
        - bearerAuth: []
      parameters:
        - name: window_days
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 180
          description: Rolling window in days (default 30)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContestStatsResponse"
        "401":
          description: Unauthorized

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
        timezone:
          type: string
          description: IANA timezone name
        min_interval_days:
          type: integer
          minimum: 1
        due_hour_local:
          type: integer
          minimum: 0
          maximum: 23
        due_minute_local:
          type: integer
          minimum: 0
          maximum: 59
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    LogoutRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
    MeResponse:
      type: object
      required: [user_id, timezone, min_interval_days, due_hour_local, due_minute_local]
      properties:
        user_id:
          type: string
        timezone:
          type: string
        min_interval_days:
          type: integer
        due_hour_local:
          type: integer
        due_minute_local:
          type: integer
    PatchSettingsRequest:
      type: object
      properties:
        timezone:
          type: string
        min_interval_days:
          type: integer
          minimum: 1
        due_hour_local:
          type: integer
          minimum: 0
          maximum: 23
        due_minute_local:
          type: integer
          minimum: 0
          maximum: 59
    SettingsResponse:
      type: object
      required: [timezone, min_interval_days, due_hour_local, due_minute_local]
      properties:
        timezone:
          type: string
        min_interval_days:
          type: integer
        due_hour_local:
          type: integer
        due_minute_local:
          type: integer
    CreateProblemRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
        title:
          type: string
        platform:
          type: string
        difficulty:
          type: string
        topics:
          type: array
          items:
            type: string
        initial_review:
          $ref: "#/components/schemas/InitialReviewRequest"
    Problem:
      type: object
      required: [id, url, platform, title, difficulty, topics]
      properties:
        id:
          type: string
        url:
          type: string
        platform:
          type: string
        title:
          type: string
        difficulty:
          type: string
        topics:
          type: array
          items:
            type: string
    UserState:
      type: object
      required: [reps, interval_days, ease, due_at, is_active]
      properties:
        reps:
          type: integer
        interval_days:
          type: integer
        ease:
          type: number
        due_at:
          type: string
          format: date-time
        is_active:
          type: boolean
    ProblemWithState:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            state:
              $ref: "#/components/schemas/UserState"
    PatchProblemRequest:
      type: object
      properties:
        is_active:
          type: boolean
        platform:
          type: string
        title:
          type: string
        difficulty:
          type: string
          description: easy|medium|hard|unknown
        topics:
          type: array
          items:
            type: string
    PostReviewRequest:
      type: object
      required: [problem_id, grade]
      properties:
        problem_id:
          type: string
        grade:
          type: integer
          minimum: 0
          maximum: 4
        time_spent_sec:
          type: integer
          minimum: 0
        source:
          type: string
        reviewed_at:
          type: string
          description: RFC3339 timestamp (optional)
    PostReviewResponse:
      type: object
      properties:
        problem_id:
          type: string
        reviewed_at:
          type: string
          format: date-time
        next_due_at:
          type: string
          format: date-time
        reps:
          type: integer
        interval_days:
          type: integer
        ease:
          type: number
        min_interval_days:
          type: integer
    RotateICSResponse:
      type: object
      required: [subscription_url]
      properties:
        subscription_url:
          type: string

    InitialReviewRequest:
      type: object
      required: [grade]
      properties:
        grade:
          type: integer
          minimum: 0
          maximum: 4
        time_spent_sec:
          type: integer
          minimum: 0
        reviewed_at:
          type: string
          description: RFC3339 timestamp (optional)
        source:
          type: string

    CreateListRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    ImportListRequest:
      type: object
      required: [template_key, version]
      properties:
        template_key:
          type: string
          description: "blind75 | neetcode150"
        version:
          type: string
          description: "v1"
    List:
      type: object
      required: [id, name, description, source_type, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        source_type:
          type: string
        source_key:
          type: string
          nullable: true
        version:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    ListItem:
      type: object
      required: [order_index, problem]
      properties:
        order_index:
          type: integer
        problem:
          $ref: "#/components/schemas/Problem"
    ListWithItems:
      allOf:
        - $ref: "#/components/schemas/List"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ListItem"
    AddListItemRequest:
      type: object
      required: [problem_id]
      properties:
        problem_id:
          type: string
    ReorderListRequest:
      type: object
      required: [problem_ids]
      properties:
        problem_ids:
          type: array
          items:
            type: string

    DifficultyMix:
      type: object
      required: [easy, medium, hard]
      properties:
        easy:
          type: integer
          minimum: 0
        medium:
          type: integer
          minimum: 0
        hard:
          type: integer
          minimum: 0
    GenerateContestRequest:
      type: object
      required: [duration_minutes, strategy, difficulty_mix]
      properties:
        duration_minutes:
          type: integer
          minimum: 1
        strategy:
          type: string
          description: "balanced | weakness | due-heavy"
        difficulty_mix:
          $ref: "#/components/schemas/DifficultyMix"
    Contest:
      type: object
      required: [id, user_id, duration_minutes, strategy, created_at]
      properties:
        id:
          type: string
        user_id:
          type: string
        duration_minutes:
          type: integer
        strategy:
          type: string
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
    ContestItem:
      type: object
      required: [order_index, target_minutes, problem]
      properties:
        order_index:
          type: integer
        target_minutes:
          type: integer
        problem:
          $ref: "#/components/schemas/Problem"
    ContestWithItems:
      allOf:
        - $ref: "#/components/schemas/Contest"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ContestItem"
    SubmitContestResult:
      type: object
      required: [problem_id, grade]
      properties:
        problem_id:
          type: string
        grade:
          type: integer
          minimum: 0
          maximum: 4
        time_spent_sec:
          type: integer
          minimum: 0
        solved_flag:
          type: boolean
    SubmitContestResultsRequest:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/SubmitContestResult"

    StatsOverview:
      type: object
      required:
        - active_problems
        - overdue_count
        - due_today_count
        - due_soon_count
        - reviews_last_7_days
        - current_streak_days
      properties:
        active_problems:
          type: integer
        overdue_count:
          type: integer
        due_today_count:
          type: integer
        due_soon_count:
          type: integer
        reviews_last_7_days:
          type: integer
        current_streak_days:
          type: integer
    TopicStat:
      type: object
      required: [topic, count, mastery_avg]
      properties:
        topic:
          type: string
        count:
          type: integer
        mastery_avg:
          type: number
    StreaksResponse:
      type: object
      required: [current_streak_days]
      properties:
        current_streak_days:
          type: integer

    ContestStatsResponse:
      type: object
      required: [window_days, totals, days, recent]
      properties:
        window_days:
          type: integer
        totals:
          $ref: "#/components/schemas/ContestStatsTotals"
        days:
          type: array
          items:
            $ref: "#/components/schemas/ContestStatsDay"
        recent:
          type: array
          items:
            $ref: "#/components/schemas/ContestStatsRecent"

    ContestStatsTotals:
      type: object
      required: [contests_finished, problems_recorded, solved_count, total_time_sec]
      properties:
        contests_finished:
          type: integer
        problems_recorded:
          type: integer
        solved_count:
          type: integer
        avg_grade:
          type: number
          format: double
          nullable: true
        total_time_sec:
          type: integer

    ContestStatsDay:
      type: object
      required: [date, contests_finished, problems_recorded, solved_count, total_time_sec]
      properties:
        date:
          type: string
          description: YYYY-MM-DD in user timezone
        contests_finished:
          type: integer
        problems_recorded:
          type: integer
        solved_count:
          type: integer
        avg_grade:
          type: number
          format: double
          nullable: true
        total_time_sec:
          type: integer

    ContestStatsRecent:
      type: object
      required:
        [
          contest_id,
          strategy,
          duration_minutes,
          created_at,
          total_items,
          recorded_count,
          solved_count,
          total_time_sec,
        ]
      properties:
        contest_id:
          type: string
        strategy:
          type: string
        duration_minutes:
          type: integer
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        total_items:
          type: integer
        recorded_count:
          type: integer
        solved_count:
          type: integer
        avg_grade:
          type: number
          format: double
          nullable: true
        total_time_sec:
          type: integer